Return type fixes
++ added new return type for Wall.GetComments method that now contains
missing fields
++ fixed return type for Wall.GetById

+++ VkLib.Core/Types/Video/Video.cs
@@ -82,7 +82,7 @@ namespace VkLib.Types.Video
         /// Date when video has been uploaded in Unixtime
         /// </summary>
         [JsonProperty("date")]
-        public int? Date { get; set; }
+        public double? Date { get; set; }
 
         /// <summary>
         /// Video owner ID
